Offline detector for car charger
