+++ src/main/resources/application.conf
@@ -197,6 +197,8 @@ processor {
         status-item = "CarChargerStatusSHS"
         power-item = "CarChargerPowerSHS"
         energy-today-item = "CarChargerEnergiaAvuiSHS"
+        online-status-item = "CarChargerMicroControladorEstatSHS"
+        offline-notification = "El dispositiu de control del carregador no respon"
         id = "car-charger-processor"
     }
 }
